Add release workflow for Github Actions
This actions compiles the system emulator with the relevant options
and creates an archive that contains this binary, as well as some OT
relevant scripts to prepare OTP/flash images. The script uses the
GitHub CLI to then create a release and publish the binary. This
automatically tags the commit as well. GitHub takes care of creating
an archive for the source code.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>