[ot] Do not drop data on PTY write if it is connected
The current code in char-pty.c is a bit odd: is uses poll to
determine when it can read from a device, but it does not poll
for write. Instead, it uses a timer to regularly determine whether
the PTY is connected. We can always determine the connection status
of a PTY by polling the HUP flag (up for disconnected, down for
connected). This commit refactors the code to get rid of the timer
and simply poll with no delay on every write to see if it connected.

A better solution would be to use an I/O watch with a callback.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>