[ot] hw/opentitan: ot_usbdev: emulate the frame counter and IRQ
Previously, the device was not properly emulating the frame counter
and would never trigger the frame IRQ. With this commit, both are
now properly emulated. There is also an optimization to avoid
triggering a timer too many times if not needed (see explanation in
the code). As a result, the device will also now properly emulate
the powered -> active_nosof -> active link state transitions.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>